requirements.txt
africastalking==1.2.5
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==5.0.6
django-ajax-selects==2.2.0
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-countries==7.6.1
django-crispy-forms==2.1
django-debug-toolbar==4.4.2
django-decouple==2.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.32.3
schema==0.7.7
sqlparse==0.5.0
stripe==9.8.0
typing_extensions==4.12.0
urllib3==2.2.1